Update installation script and create uninstall script
The installation script now checks for a previously linked copy of
yeoman-foundation.  If the symlinked foundation directory exists in the
global yeoman generators directory, it will be unlinked and then a new
symlink will be created (not sure if the unlink then relink step is
really necessary for symbolic links but it won't hurt).

An uninstall script has been created to clean up after itself if a user
decides to uninstall this package.  `package.json` now has a entry for
`uninstall` that will run `uninstall.js` when a user uninstalls via
`[sudo] npm uninstall -g yeoman-foundation`.  `uninstall.js` simply
removes the symlink in the global yeoman generators directory before
the package is uninstalled (see: https://npmjs.org/doc/scripts.html)